boocking_service-main/requirements.txt
Django==4.1.8
gunicorn==20.1.0
django-cors-headers
whitenoise==6.4.0
dj_database_url==1.3.0
psycopg2-binary==2.9.6
django-environ==0.10.0
djongo==1.3.6
pytz
ipython
pymongo~=4.3.3
djangorestframework==3.14.0
drf-spectacular
djangorestframework-simplejwt
channels[daphne]~=4.0.0
channels_redis

asgiref~=3.6.0
pandas~=2.0.1
numpy~=1.24.3
requests~=2.30.0
urllib3==1.26.6
scikit-learn~=1.2.2
openpyxl